poc for read only project
