+++ src/neptune/objects/neptune_object.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-__all__ = ["NeptuneObject"]
+__all__ = ["NeptuneObject", "WithBackend"]
 
 import abc
 import atexit
@@ -36,7 +36,6 @@ from typing import (
     Any,
     Callable,
     Dict,
-    Iterable,
     List,
     Optional,
     Union,
@@ -70,7 +69,6 @@ from neptune.handler import Handler
 from neptune.internal.backends.api_model import Project
 from neptune.internal.backends.factory import get_backend
 from neptune.internal.backends.neptune_backend import NeptuneBackend
-from neptune.internal.backends.nql import NQLQuery
 from neptune.internal.backends.project_name_lookup import project_name_lookup
 from neptune.internal.backgroud_job_list import BackgroundJobList
 from neptune.internal.background_job import BackgroundJob
@@ -108,9 +106,7 @@ from neptune.internal.utils.utils import reset_internal_ssl_state
 from neptune.internal.value_to_attribute_visitor import ValueToAttributeVisitor
 from neptune.internal.warnings import warn_about_unsupported_type
 from neptune.objects.mode import Mode
-from neptune.table import Table
 from neptune.types.type_casting import cast_value
-from neptune.typing import ProgressBarType
 from neptune.utils import stop_synchronization_callback
 
 if TYPE_CHECKING:
@@ -142,8 +138,47 @@ def temporarily_disabled(func: Callable[P, R]) -> Callable[P, R]:
     return wrapper
 
 
-class NeptuneObject(AbstractContextManager):
+class WithBackend(AbstractContextManager):
     container_type: ContainerType
+
+    def __init__(
+        self,
+        api_token: Optional[str] = None,
+        project: Optional[str] = None,
+        mode: Mode = Mode.ASYNC,
+        proxies: Optional[dict] = None,
+    ) -> None:
+        verify_type("api_token", api_token, (str, type(None)))
+        verify_type("mode", mode, Mode)
+        verify_type("proxies", proxies, (dict, type(None)))
+        verify_type("project", project, (str, type(None)))
+
+        self._mode = mode
+        self._backend: NeptuneBackend = get_backend(mode=mode, api_token=api_token, proxies=proxies)
+        self._project_qualified_name: Optional[str] = conform_optional(project, QualifiedName)
+        self._project_api_object: Project = project_name_lookup(
+            backend=self._backend, name=self._project_qualified_name
+        )
+        self._workspace: str = self._project_api_object.workspace
+        self._project_name: str = self._project_api_object.name
+        self._project_id: UniqueId = self._project_api_object.id
+
+        self._state: ContainerState = ContainerState.CREATED
+
+    def stop(self, *, seconds: Optional[Union[float, int]] = None) -> None:
+        verify_type("seconds", seconds, (float, int, type(None)))
+
+        if self._state != ContainerState.STARTED:
+            return
+        self._backend.close()
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if exc_tb is not None:
+            traceback.print_exception(exc_type, exc_val, exc_tb)
+        self.stop()
+
+
+class NeptuneObject(WithBackend):
 
     def __init__(
         self,
@@ -160,11 +195,7 @@ class NeptuneObject(AbstractContextManager):
         async_no_progress_threshold: float = ASYNC_NO_PROGRESS_THRESHOLD,
     ):
         verify_type("custom_id", custom_id, (str, type(None)))
-        verify_type("project", project, (str, type(None)))
-        verify_type("api_token", api_token, (str, type(None)))
-        verify_type("mode", mode, Mode)
         verify_type("flush_period", flush_period, (int, float))
-        verify_type("proxies", proxies, (dict, type(None)))
         verify_type("async_lag_threshold", async_lag_threshold, (int, float))
         verify_optional_callable("async_lag_callback", async_lag_callback)
         verify_type("async_no_progress_threshold", async_no_progress_threshold, (int, float))
@@ -176,24 +207,15 @@ class NeptuneObject(AbstractContextManager):
                 f"Ensure that the `custom_run_id` argument doesn't exceed the limit."
             )
 
-        self._mode: Mode = mode
+        super().__init__(api_token=api_token, project=project, mode=mode, proxies=proxies)
+
         self._flush_period = flush_period
         self._lock: threading.RLock = threading.RLock()
         self._forking_cond: threading.Condition = threading.Condition()
         self._forking_state: bool = False
-        self._state: ContainerState = ContainerState.CREATED
         self._signals_queue: "Queue[Signal]" = Queue()
         self._logger: logging.Logger = get_logger()
 
-        self._backend: NeptuneBackend = get_backend(mode=mode, api_token=api_token, proxies=proxies)
-
-        self._project_qualified_name: Optional[str] = conform_optional(project, QualifiedName)
-        self._project_api_object: Project = project_name_lookup(
-            backend=self._backend, name=self._project_qualified_name
-        )
-        self._workspace: str = self._project_api_object.workspace
-        self._project_name: str = self._project_api_object.name
-        self._project_id: UniqueId = self._project_api_object.id
         self._custom_id: str = custom_id or str(uuid.uuid4())
 
         self._async_lag_threshold = async_lag_threshold
@@ -337,11 +359,6 @@ class NeptuneObject(AbstractContextManager):
 
     def _write_initial_attributes(self):
         pass
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if exc_tb is not None:
-            traceback.print_exception(exc_type, exc_val, exc_tb)
-        self.stop()
 
     def __getattr__(self, item):
         raise AttributeError(f"'{self.__class__.__name__}' object has no attribute '{item}'")
@@ -682,39 +699,6 @@ class NeptuneObject(AbstractContextManager):
 
     def _shutdown_hook(self):
         self.stop()
-
-    def _fetch_entries(
-        self,
-        child_type: ContainerType,
-        query: NQLQuery,
-        columns: Optional[Iterable[str]],
-        limit: Optional[int],
-        sort_by: str,
-        ascending: bool,
-        progress_bar: Optional[ProgressBarType],
-    ) -> Table:
-        if columns is not None:
-            # always return entries with 'sys/id' and the column chosen for sorting when filter applied
-            columns = set(columns)
-            columns.add("sys/id")
-            columns.add(sort_by)
-
-        leaderboard_entries = self._backend.search_leaderboard_entries(
-            project_id=self._project_id,
-            types=[child_type],
-            query=query,
-            columns=columns,
-            limit=limit,
-            sort_by=sort_by,
-            ascending=ascending,
-            progress_bar=progress_bar,
-        )
-
-        return Table(
-            backend=self._backend,
-            container_type=child_type,
-            entries=leaderboard_entries,
-        )
 
     def get_root_object(self) -> "NeptuneObject":
         """Returns the same Neptune object."""
